requirements.txt
apache-airflow==2.8.4
psycopg2-binary==2.9.9
boto3==1.35.35
pandas==2.2.3
